Move GtkStore contents (WindowStore) into OutputMethods
